+++ app/src/main/java/one/mann/weatherman/model/openWeatherMap/Main.kt
@@ -2,8 +2,8 @@ package one.mann.weatherman.model.openWeatherMap
 
 class Main {
     val temp: Float = 0.toFloat()
-    val pressure: Int = 0
-    val humidity: Int = 0
-    val temp_min: Int = 0
-    val temp_max: Int = 0
+    val pressure: Float = 0.toFloat()
+    val humidity: Float = 0.toFloat()
+    val temp_min: Float = 0.toFloat()
+    val temp_max: Float = 0.toFloat()
 }
